Use tempfile to read from stdin
This can be achieved by giving "-" as a filename. Internally, we create
a temporary file, write the contents of stdin to it and read from it on
each iteration. This way, we can use input from stdin even in daemon
mode, where file handles to stdin, stdout, and stderr are closed.